Calculate AnthropoAgeAccel to obtain a metric of biological age acceleration

anthropoage_accel() is used to estimate AnthropoAgeAccel, which are the sex-stratified residuals of the regression of Age onto AnthropoAge. This function does not distinguish between AnthropoAge and S-AnthropoAge and it may be used interchangebly for either metric. Computation is recommended in databases of at least 10 individuals and its performance will be better for larger datasets. The function does not calculate age acceleration for datasets with n<10.

Estimation of AnthropopoAgeAccel uses as inputs chronological age (Age), sex (Sex, codified as "Men" or "Women"), and AnthropoAge (either the complete or simplified version).

AnthropoAgeAccel is interpreted as the average deviation of biological from chronological age and is specific for each population. Values ≥0 indicate accelerated aging rates, whilst values<0 indicate non-accelerated aging rates.

anthropoage_accel(Age, Sex, AnthropoAge)

Arguments

Age

Chronological age of an individual expressed in full years.

Sex

Sex of an individual, it can either take the value "Men" or "Women". The function does not recognize numeric values.

AnthropoAge

Proxy of biological age, measured in years. This values should be estimated using the anthropoage( ) or s_anthropoage( ) functions within this package.

Details

Value

References

Author

Dr. Omar Yaxmehen Belo Chavolla & Dr. Carlos Alberto Fermín Martínez

Bello-Chavolla Lab, November 2023

Full paper here: https://onlinelibrary.wiley.com/doi/10.1111/acel.13756

Fermín-Martínez CA, Márquez-Salinas A, Guerra EC, Zavala-Romero L, Antonio-Villa NE, Fernández-Chirino L, Sandoval-Colin E, Barquera-Guevara DA, Campos Muñoz A, Vargas-Vázquez A, Paz-Cabrera CD, Ramírez-García D, Gutiérrez-Robledo LM, Bello-Chavolla OY. AnthropoAge, a novel approach to integrate body composition into the estimation of biological age. Aging Cell. 2023 Jan;22(1):e13756. doi: 10.1111/acel.13756. Epub 2022 Dec 22. PMID: 36547004; PMCID: PMC9835580.

Note

For datasets with n<10 a simple substraction of Chronological age minus AnthropoAge may be used as a proxy of age acceleration. This, however, would not consider deviation from population average rates.

See also

Examples

## Calculate s_anthropoage in a single individual ##

accel<-anthropoage_accel(Age=31, AnthropoAge=26.2, Sex="Men")

accel
#> [1] "For optimal calculations please only estimate AntropoAgeAccel for n≥10"